+++ packages/gamut/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [68.1.0](https://github.com/Codecademy/gamut/compare/@codecademy/gamut@68.0.5...@codecademy/gamut@68.1.0) (2026-03-12)
+
+### Features
+
+- Logical properties ([4c340c3](https://github.com/Codecademy/gamut/commit/4c340c3949bad30fce4f9ffabb2018763df5c4ba))
+
+### [68.0.5](https://github.com/Codecademy/gamut/compare/@codecademy/gamut@68.0.4...@codecademy/gamut@68.0.5) (2026-03-04)
+
+### Bug Fixes
+
+- **pkgs:** downgrade react-aria components for CSP ([9bb66e8](https://github.com/Codecademy/gamut/commit/9bb66e8f9960e2564c91d8c9cf682b5424346c2c))
+
 ### [68.0.4](https://github.com/Codecademy/gamut/compare/@codecademy/gamut@68.0.3...@codecademy/gamut@68.0.4) (2026-03-02)
 
 ### Bug Fixes